unit tests / bugfixes for array stuff
Wrote some unit tests for the Array stuff, and lo and behold it revealed
some bugs. Neat! Those bugs have been fixed.. mostly.

I'm not quite sure why the append-and-sort test crashes - but _only_
with isize=20? I'm guessing that we reallocate memory and Something
Happens and it goes kaboom. But it's 5pm on a Friday so this is enough
for now.